requirements.txt
PyQt6>=6.5.0
pypdf>=6.0.0
pdf2image>=1.17.0
Pillow>=10.0.0
pyinstaller>=6.0.0
